Extract codex-plugin crate
Move shared plugin identifiers, capability summaries, and load outcome types into a dedicated crate while keeping codex-core as the integration layer.
