+++ codex-rs/core/src/plugins/store_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use codex_plugin::PluginId;
 use pretty_assertions::assert_eq;
 use tempfile::tempdir;
